refactor(svg): #22 moved more svgs to separate files
